Fix : Fix Front Structure
