+++ src/main/java/kr/co/monitoringserver/controller/api/AdminApiController.java
@@ -5,14 +5,12 @@ import kr.co.monitoringserver.service.dtos.request.AdminReqDTO;
 import kr.co.monitoringserver.service.dtos.request.UserReqDTO;
 import kr.co.monitoringserver.service.dtos.response.ResponseDto;
 import kr.co.monitoringserver.service.service.AdminService;
+import kr.co.monitoringserver.service.service.UserService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.validation.BindingResult;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.PutMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 
 import java.util.Map;
 @RestController
@@ -21,6 +19,24 @@ public class AdminApiController {
 
     @Autowired
     private AdminService adminService;
+
+    @Autowired
+    private UserService userService;
+
+    /**
+     * saveUser : 사용자정보를 Create하여 회원가입을 수행한다.
+     */
+    @PostMapping("/admin/joinProc")
+    public ResponseDto<?> saveUser(@Valid @RequestBody UserReqDTO userDto, BindingResult bindingResult) {
+        if(bindingResult.hasErrors()) {
+            Map<String, String> validatorResult = userService.validateHandling(bindingResult);
+
+            return new ResponseDto<>(HttpStatus.BAD_REQUEST.value(), validatorResult);
+        }
+
+        userService.join(userDto);
+        return new ResponseDto<Integer>(HttpStatus.OK.value(), 1);
+    }
 
     /**
      * updateUser : 회원정보를 Update 한다.
