+++ src/main/resources/static/js/attendance/attendance.js
@@ -1,54 +1,11 @@
 let index = {
 		init: function() {
-			$("#btn-register").on("click", ()=>{
-				this.register();
-			});
             $("#btn-update").on("click", ()=>{
                 this.update();
             });
             $("#btn-del").on("click", ()=>{
                 this.del();
             });
-		},
-
-		register: function() {
-		    var param1 = $("#userId").val();
-            var param2 = $("#userIdentity").val();
-
-            var today = new Date();
-            var year = today.getFullYear();
-            var month = ('0' + (today.getMonth() + 1)).slice(-2);
-            var day = ('0' + today.getDate()).slice(-2);
-            var dateString = year + '-' + month  + '-' + day;
-
-            var hours = ('0' + today.getHours()).slice(-2);
-            var minutes = ('0' + today.getMinutes()).slice(-2);
-            var seconds = ('0' + today.getSeconds()).slice(-2);
-            var timeString = hours + ':' + minutes  + ':' + seconds;
-
-			let data = {
-					enterTime: timeString,
-					leaveTime: timeString,
-					date: dateString
-			};
-
-			$.ajax({
-				type: "POST",
-				url: "/api/v1/attendance/"+param2,
-				data: JSON.stringify(data),
-				contentType: "application/json; charset=utf-8",
-				dataType: "json"
-			}).done(function(resp) {
-			    if(resp.status == 400 | resp.status == 500){
-                    alert("출근처리에 실패하였습니다.");
-                }
-                else{
-                    alert("출근처리가 완료되었습니다.");
-                    location.href = "/attendance/list/"+param1;
-			    }
-			}).fail(function(error) {
-				alert(JSON.stringify(error));
-			});
 		},
 
 		update: function() {
