+++ src/main/webapp/WEB-INF/views/admin/attendance/updateForm.jsp
@@ -7,20 +7,20 @@
         <div class="form-group">
             <input type="hidden" class="form-control" value="${list.user.userId}" id="userId">
             <div class="form-group">
-                <label for="identity">Id</label>
+                <label for="identity">사용자 ID</label>
                 <input type="identity" class="form-control" value="${list.user.identity}" id="identity" readonly>
             </div>
         </div>
         <div class="form-group">
-            <label for="date">Date</label>
+            <label for="date">출석일자</label>
             <input type="date" class="form-control" value="${list.attendance.date}" id="date" readonly>
         </div>
 		<div class="form-group">
-			<label for="enterTime">EnterTime</label>
+			<label for="enterTime">출근시간</label>
 			<input type="time" class="form-control" id="enterTime">
 		</div>
         <div class="form-group">
-            <label for="leaveTime">LeaveTime</label>
+            <label for="leaveTime">퇴근시간</label>
             <input type="time" class="form-control" id="leaveTime">
         </div>
 	</form>
