+++ src/main/webapp/WEB-INF/views/admin/securityArea/detail.jsp
@@ -6,19 +6,19 @@
 	<form>
             <input type="hidden" value="${list.id}" class="form-control" id="id">
 		<div class="form-group">
-			<label for="name">Id</label>
+			<label for="name">보안구역명</label>
 			<input type="text" value="${list.name}" class="form-control" placeholder="Enter name" id="name">
 		</div>
 		<div class="form-group">
-			<label for="description">Description</label>
+			<label for="description">명세</label>
 			<input type="description" value="${list.description}" class="form-control" placeholder="Enter description" id="description">
 		</div>
 		<div class="form-group">
-			<label for="latitude">Latitude</label>
+			<label for="latitude">위도</label>
 			<input type="latitude" value="${list.location.latitude}" class="form-control" placeholder="Enter latitude" id="latitude">
 		</div>
 		<div class="form-group">
-            <label for="longitude">Longitude</label>
+            <label for="longitude">경도</label>
             <input type="longitude" value="${list.location.longitude}" class="form-control" placeholder="Enter longitude" id="longitude">
         </div>
 	</form>
