+++ src/main/webapp/WEB-INF/views/index.jsp
@@ -12,14 +12,12 @@
 
           <h5>🛑 주의사항</h5>
             <h6>- 출근 시 퇴근할 때까지 스마트폰의 블루투스를 켜둔 상태로 유지합니다.</h6>
-            <h6>- 위 사항을 지키지 않을 경우, 보안문제 발생 시 불리하게 작용합니다.</h6>
+            <h6>- 위 사항을 지키지 않을 경우, 보안문제 발생 시 불리하게 작용됩니다.</h6>
 
             <br>
 
           <h5>✅ 출결관리</h5>
-            <h6>- 출결조회 : 사용자의 출근 및 결근 기록을 조회합니다.</h6>
-            <h6>- 출결등록 : 사용자의 출근 및 퇴근 여부를 등록합니다.</h6>
-            <h6>- 출결현황 : 사용자의 출결 현황을 조회합니다.</h6>
+            <h6>- 출결조회 : 사용자의 출결 기록 및 현황을 조회합니다.</h6>
 
             <br>
