+++ src/main/webapp/WEB-INF/views/user/attendance/condition.jsp
@@ -27,6 +27,23 @@
             <input type="text" value="${absent}" class="form-control" id="absent" readonly>
         </div>
 	</form>
+	<button id="btn-back" class="btn btn-dark">뒤로가기</button>
 </div>
+
+<script>
+    let index = {
+        init: function() {
+                    $("#btn-back").on("click", ()=>{
+                        this.back();
+                    });
+                },
+
+                back: function() {
+                    window.history.back();
+                }
+        }
+    index.init();
+</script>
+
 
 <%@ include file="../../layout/user/footer.jsp"%>
