+++ src/main/webapp/WEB-INF/views/user/attendance/inquire.jsp
@@ -10,17 +10,16 @@
   <form style="text-align:right;" action="${principal.user.userId}" method="get">
     <input type="date" name="searchKeyword">
     <button type="submit" class="btn btn-dark">검색</button>
-  <form>
   <table class="table table-hover">
     <thead>
       <tr>
-        <th>Date</th>
-        <th>Name</th>
-        <th>Department</th>
-        <th>EnterTime</th>
-        <th>LeaveTime</th>
-        <th>GoWork</th>
-        <th>LeaveWork</th>
+        <th>날짜</th>
+        <th>이름</th>
+        <th>부서</th>
+        <th>출근시간</th>
+        <th>퇴근시간</th>
+        <th>지각여부</th>
+        <th>조퇴여부</th>
       </tr>
     </thead>
     <tbody>
@@ -38,7 +37,9 @@
     </tbody>
   </table>
 
-  <ul class="pagination justify-content-center">
+  <button style="display:inline-block; float:left;" type="button" id="btn-condition" class="btn btn-dark">출결현황조회</button>
+
+  <ul class="pagination" style="position:relative; left:35%">
   	<c:choose>
   		<c:when test="${lists.first}">
   			<li class="page-item disabled"><a class="page-link" href="?page=${lists.number-1}">Previous</a></li>
@@ -66,8 +67,23 @@
   		<c:otherwise>
   			<li class="page-item"><a class="page-link" href="?page=${lists.number+1}">Next</a></li>
   		</c:otherwise>
-  	</c:choose>
+  	    </c:choose>
+  	</ul>
   </form>
 </div>
+<script>
+    let index = {
+        init: function() {
+                    $("#btn-condition").on("click", ()=>{
+                        this.condition();
+                    });
+                },
+
+                condition: function() {
+                    location.href="/attendance/list/condition/${principal.user.userId}"
+                }
+        }
+    index.init();
+</script>
 
 <%@ include file="../../layout/user/footer.jsp"%>
