+++ src/main/webapp/WEB-INF/views/user/loginForm.jsp
@@ -5,11 +5,11 @@
 <div class="container" >
 	<form action="/auth/loginProc" method="POST">
 		<div class="form-group">
-			<label for="identity">Id</label>
+			<label for="identity">ID</label>
 			<input type="text" name="identity" class="form-control" placeholder="Enter identity" id="identity">
 		</div>
 		<div class="form-group">
-			<label for="pwd">Password</label> 
+			<label for="pwd">패스워드</label>
 			<input type="password" name="password" class="form-control" placeholder="Enter password" id="password">
 		</div>
 			<button id="btn-login" class="btn btn-dark">로그인</button>
